Create a-history-of-node.md
This is a "Talk Starter" file. Much like https://github.com/cascadiajs/2015.cascadiajs.com/issues/4, it occurred to me "starter ideas" might help drive talk submissions. These starters would include a rough draft of what the community might like to see in a talk, and prospective speakers could expand the starter into their own talks. Because the very idea of the talk is "up for grabs", we would expect proposals based on off these ideas to be very well defined indeed.

This is a new file (rather than an issue) to enable both a deeper revision history and a super-obvious "best state" of the starter idea. If/when this format for starts is approved, I will add some other starters I have.